Use Spectral Storage service, Remove Proxl DB tables
--  Drop table scan_spectrum_data since no longer used
DROP TABLE scan_spectrum_data;
--  remove all records in scan table with level 1
DELETE FROM scan WHERE level = 1
DROP TABLE scan_file_ms1_intensity_binned_summed_data;
DROP TABLE scan_retention_time;

+++ proxl_web_app/src/org/yeastrc/xlink/www/lorikeet_dto/LorikeetRootData.java
@@ -82,7 +82,7 @@ public class LorikeetRootData {
 //	jpeak.add( peak.getMz() );
 //	jpeak.add( peak.getIntensity() );
 //
-//	    private double mz;
+//	    private float mz;
 //    private float intensity;
 
     /**
@@ -90,7 +90,7 @@ public class LorikeetRootData {
      * @param mz
      * @param intensity
      */
-    public void addPeak( double mz, float intensity ) {
+    public void addPeak( float mz, float intensity ) {
     	
     	if ( peaks == null ) {
     		
@@ -111,7 +111,7 @@ public class LorikeetRootData {
      * @param mz
      * @param intensity
      */
-    public void addMs1Peak( double mz, float intensity ) {
+    public void addMs1Peak( float mz, float intensity ) {
     	
     	if ( ms1peaks == null ) {
